This is the first commit for some time. 1. Removed experimental 2. Added RopeScene 3. Added RCRopeEx2.project in prep for RopeScene 4. Improved JSON input 5. Moved sln to VS 2022 6. Minor rework of InputAction. Still undecided on inheriting enum BaseUserCommand 7. GeometricShapes - moved from float to double 8. Added some shaders

+++ engine/Helpers/GeometricShapes.cpp
@@ -100,10 +100,10 @@ std::vector<glm::vec3> GeometricShapes::star(float innerRadius, float outerRadiu
 								unsigned int numPoints)
 {
 	std::vector<glm::vec3> retval;
-	const float incr = glm::two_pi<float>() / numPoints;
-	float angle1 = 0;
-	float angle2 = angle1 + incr / 2.0;
-	float angle3 = angle2 + incr / 2.0;
+	const double incr = glm::two_pi<float>() / numPoints;
+	double angle1 = 0;
+	double angle2 = angle1 + incr / 2.0;
+	double angle3 = angle2 + incr / 2.0;
 
 	for (unsigned int i = 0; i < numPoints; i++)
 	{
